Make annotation macro derive Canon
In order to have Generic and easy-to-implement
`Anotations` which use Canon-derived structures inside, we needed
to have the `Canon` trait implemented for them.

Now the macro derives the Canon trait correctly and it is working
for poseidon252::PoseidonAnnotation which is generated using this
exact macro.

Also the implementation of `Content` has been removed since the
trait has been deprecated.